fix(review): resolve the independent review's findings
No majors were found. Ten minor findings, all fixed:

- A hand-off in the event that opened a batch let exactly one queued event run
  behind the child, because the flag was checked after handling rather than
  before. Both call sites now ask one tested rule (batch_step), so the entry
  guard and the in-loop check cannot disagree.
- scroll-lines above 255 failed to parse as a u8 and took the whole config
  file down with it, clamping only below that. The key is read as the widest
  integer TOML carries, so 300 clamps to 10 and the other keys survive.
- A status message arriving mid-batch could be cleared by a later event in the
  same batch and never drawn, which deferring the frame was not supposed to
  change. A message now survives to the frame that shows it.
- thumb() returned a full-height thumb for a one-row track, claiming the top
  and the bottom at once; it returns None there, as it does when rows fit.
- The selection wash ran edge to edge and painted over the scrollbar column,
  erasing the track on the selected row in themes where the guide colour
  equals the wash.
- FlatView::scroll_by had no caller left while app duplicated its clamp; it
  now takes the row count and is the single place the clamp lives.
- The ADR 0015 doc comment had been separated from resolve_click and still
  described the chevron as acting on the press.
- Tests added for the paths that had none: the socket scrollbar key, the
  config scrollbar branch, arming from a hit (including an out-of-range
  index), the batch rule, a pane too narrow for the furniture, and the wash
  stopping before the bar.
- README's ctl set key list and the CHANGELOG's Unreleased section were
  missing this sprint's four user-visible changes.

The review also found a pre-existing bug outside the diff — perform_open's
stale-event drain discards handle()'s quit flag, so SIGHUP or ctl quit during
a terminal handover is answered and then ignored. Filed as 077, not fixed
here.

+++ crates/birch-core/src/settings.rs
@@ -97,9 +97,11 @@ pub const SCROLL_LINES_DEFAULT: u8 = 3;
 
 /// Clamps into the accepted range. Used where a value must degrade rather
 /// than fail (the config file); surfaces that can report an error reject
-/// instead.
-pub fn clamp_scroll_lines(n: u8) -> u8 {
-    n.clamp(SCROLL_LINES_MIN, SCROLL_LINES_MAX)
+/// instead. Takes the widest integer TOML can carry, so that a number far
+/// outside the range still clamps instead of failing to parse and taking the
+/// rest of the config file down with it.
+pub fn clamp_scroll_lines(n: i64) -> u8 {
+    n.clamp(i64::from(SCROLL_LINES_MIN), i64::from(SCROLL_LINES_MAX)) as u8
 }
 
 impl Default for Settings {
